requirements.txt
Django==2.2.2
python-aiml==0.9.1
pytz==2019.1
sqlparse==0.3.0
